Create UserObjectBase #30655
